refactor: review -- unify closely related classes
